pyproject.toml
[project]
name = "voice-agent-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg>=1.4",
    "numpy>=2.3.2",
    "openai-agents[voice]>=0.2.4",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scipy>=1.16.1",
    "sounddevice>=0.5.2",
    "speechrecognition>=3.14.3",
    "streamlit>=1.48.0",
    "streamlit-webrtc>=0.63.4",
]
